+++ catalyst/metrics/functional/_hitrate.py
@@ -5,9 +5,12 @@ import torch
 from catalyst.metrics.functional._misc import process_recsys_components
 
 
-def hitrate(outputs: torch.Tensor, targets: torch.Tensor, topk: List[int]) -> List[torch.Tensor]:
+def hitrate(
+    outputs: torch.Tensor, targets: torch.Tensor, topk: List[int], zero_division: int = 0
+) -> List[torch.Tensor]:
     """
-    Calculate the hit rate score given model outputs and targets.
+    Calculate the hit rate (aka recall) score given
+    model outputs and targets.
     Hit-rate is a metric for evaluating ranking systems.
     Generate top-N recommendations and if one of the recommendation is
     actually what user has rated, you consider that a hit.
@@ -30,6 +33,9 @@ def hitrate(outputs: torch.Tensor, targets: torch.Tensor, topk: List[int]) -> Li
             ground truth, labels
         topk (List[int]):
             Parameter fro evaluation on top-k items
+        zero_division (int):
+            value, returns in the case of the divison by zero
+            should be one of 0 or 1
 
     Returns:
         hitrate_at_k (List[torch.Tensor]): the hitrate score
@@ -39,7 +45,8 @@ def hitrate(outputs: torch.Tensor, targets: torch.Tensor, topk: List[int]) -> Li
     targets_sort_by_outputs = process_recsys_components(outputs, targets)
     for k in topk:
         k = min(outputs.size(1), k)
-        hits_score = torch.sum(targets_sort_by_outputs[:, :k], dim=1) / k
+        hits_score = torch.sum(targets_sort_by_outputs[:, :k], dim=1) / targets.sum(dim=1)
+        hits_score = hits_score.nan_to_num(zero_division)
         results.append(torch.mean(hits_score))
 
     return results